% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adev.R
\name{Adev}
\alias{Adev}
\title{Calculate the mean element-wise absolute deviation between two matrices.}
\usage{
Adev(mat_A, mat_B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix of same dimensions as \code{A}}
}
\value{
mean absolute deviation
}
\description{
Calculate the mean element-wise absolute deviation between two matrices.
}
