% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjMat.R
\name{adjMat}
\alias{adjMat}
\title{Find posterior mean of \eqn{\Sigma} given data matrix, \eqn{x}, prior correlation matrix,
\eqn{\Delta}, and weight, \eqn{\alpha}.}
\usage{
adjMat(x, Delta, alpha)
}
\arguments{
\item{x}{standardized data matrix of dimensions \eqn{n \times p}}

\item{Delta}{prior correlation matrix of dimensions \eqn{p \times p} (positive semi-definite)}

\item{alpha}{weight to use in average between sample correlation matrix and prior correlation matrix}
}
\value{
An object of class \sQuote{data.frame}
}
\description{
Find posterior mean of \eqn{\Sigma} given data matrix, \eqn{x}, prior correlation matrix,
\eqn{\Delta}, and weight, \eqn{\alpha}.
}
