% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustSign.R
\name{AdjustSign}
\alias{AdjustSign}
\title{Perform a sign operation on matrix \eqn{\Sigma} according to the sign of \eqn{A_I}.}
\usage{
AdjustSign(Sigma, AI)
}
\arguments{
\item{Sigma}{a sample correlation matrix of dimensions \eqn{p \times p}}

\item{AI}{a matrix of dimensions \eqn{p \times K}}
}
\value{
a matrix of dimensions \eqn{p \times p}
}
\description{
Perform a sign operation on matrix \eqn{\Sigma} according to the sign of \eqn{A_I}.
}
