% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBMA.R
\name{betaBMA}
\alias{betaBMA}
\title{PriorER - Bayesian Model Averaging For \eqn{\beta}s}
\usage{
betaBMA(x, y, er_res, imps)
}
\arguments{
\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{y}{a response vector of dimension \eqn{n}}

\item{er_res}{the result of \link{plainER} or \link{priorER}}

\item{imps}{a vector of important variables}
}
\value{
nothing is returned, saves boxplot of cross-validation results for user to use
in selecting optimal \eqn{\delta}
}
\description{
Find new estimates for \eqn{\beta}s using Bayesian Adaptive
Sampling for Bayesian Model Averaging as in the \code{BAS} package by Clyde et al.
}
