% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calArates.R
\name{calArates}
\alias{calArates}
\title{Calculate the scaled matrix \eqn{L_1} and Frobenius norms.}
\usage{
calArates(A)
}
\arguments{
\item{A}{a matrix}
}
\value{
a vector containing the scaled \eqn{L_1} and Frobenius norms
}
\description{
Calculate the scaled matrix \eqn{L_1} and Frobenius norms.
}
