% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalFittedSigma.R
\name{CalFittedSigma}
\alias{CalFittedSigma}
\title{Calculate the fitted value of \eqn{A_I \cdot C \cdot A_I^\top} for given
\eqn{\Sigma} and \eqn{\delta}.}
\usage{
CalFittedSigma(Sigma, delta, Ms, arg_Ms, se_est, diagonal, merge)
}
\arguments{
\item{Sigma}{a correlation matrix of dimensions \eqn{p \times p}}

\item{delta}{a threshold parameter}

\item{Ms}{the calculated maximal values of \eqn{\Sigma} by row}

\item{se_est}{the estimated standard errors}

\item{diagonal}{a boolean indicating the diagonal structure}

\item{merge}{a boolean indicating the merge style}
}
\value{
a list containing a vector of the indices of the estimated pure variables and
the fitted value of \eqn{A_I \cdot C \cdot A_I^\top}
}
\description{
Calculate the fitted value of \eqn{A_I \cdot C \cdot A_I^\top} for given
\eqn{\Sigma} and \eqn{\delta}.
}
