% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calFNRandFPR.R
\name{calFNRandFPR}
\alias{calFNRandFPR}
\title{Calculate the FNR and FPR for the estimated matrix, \eqn{A}.}
\usage{
calFNRandFPR(estA, trueA)
}
\arguments{
\item{estA}{an estimated matrix}

\item{trueA}{the true matrix of the same dimensions as \code{estA}}
}
\value{
a list containing the FNRs and FPRs as well as the row indices of where the errors occur
}
\description{
Calculate the FNR and FPR for the estimated matrix, \eqn{A}.
}
