% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calSPandSN.R
\name{calSPandSN}
\alias{calSPandSN}
\title{Calculate specificity and sensitivity based on group partition}
\usage{
calSPandSN(p, estGroup, trueGroup)
}
\arguments{
\item{p}{total number of variables}

\item{estGroup}{list of estimated variable indices}

\item{trueGroup}{list of true variable indices}
}
\value{
the specificity and sensitivity
}
\description{
Calculate specificity and sensitivity based on group partition
}
