% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comp_ASV.R
\name{Comp_ASV}
\alias{Comp_ASV}
\title{I am not sure what this does}
\usage{
Comp_ASV(sigma2_hat, BI, R, Gamma_hat, beta_hat, Omega_hat, I_hat, I_ind_hat)
}
\arguments{
\item{sigma2_hat}{???}

\item{BI}{???}

\item{R}{???}

\item{Gamma_hat}{estimated matrix}

\item{beta_hat}{matrix of the estimated values of \est{\beta}}

\item{Omega_hat}{estimated matrix}

\item{I_hat}{estimated matrix}

\item{I_ind_hat}{estimated matrix}
}
\value{
???
}
\description{
I am not sure what this does
}
