% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Delta.R
\name{CV_Delta}
\alias{CV_Delta}
\title{Cross validation for choosing \eqn{\delta}. For each delta from the given grids,
first split the data into two data sets. Obtain \eqn{I}, \eqn{A_I} and \eqn{C} from data set 1.
Then calculate \eqn{A_I \cdot C \cdot A_I^\top} and choose \eqn{\delta} which minimizes the criterion
\eqn{A_I \cdot C \cdot A_I^\top - \Sigma(\text{data set 2})}}
\usage{
CV_Delta(x, deltaGrids, diagonal, se_est, merge)
}
\arguments{
\item{x}{data matrix of dimensions \eqn{n \times p}}

\item{deltaGrids}{a vector of numerical constants over which to perform the search for the optimal \eqn{\delta}}

\item{diagonal}{a boolean indicating the diagonal structure of \eqn{C}}

\item{se_est}{estimated standard errors}

\item{merge}{a boolean indicating merge style}
}
\value{
the selected optimal \eqn{\delta}
}
\description{
Cross validation for choosing \eqn{\delta}. For each delta from the given grids,
first split the data into two data sets. Obtain \eqn{I}, \eqn{A_I} and \eqn{C} from data set 1.
Then calculate \eqn{A_I \cdot C \cdot A_I^\top} and choose \eqn{\delta} which minimizes the criterion
\eqn{A_I \cdot C \cdot A_I^\top - \Sigma(\text{data set 2})}
}
