% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_lbd.R
\name{CV_lbd}
\alias{CV_lbd}
\title{Use cross-validation to select \eqn{\lambda} for estimating \eqn{\Omega}.
Split the data into two parts and estimate \eqn{C} on both data sets. Then, for each
\eqn{lambda}, calculate \eqn{\Omega} on the first dataset and calculate the loss on the second dataset.
Find the lambda which minimizes \eqn{C \cdot \Omega - log(|\Omega|)}}
\usage{
CV_lbd(x, lbdGrids, AI, pureVec, diagonal, k)
}
\arguments{
\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{lbdGrids}{a vector of numerical constants over which to search for the optimal \eqn{\lambda}}

\item{AI}{the estimated matrix \eqn{A_I} of dimensions \eqn{p \times K}}

\item{pureVec}{a vector of indices of pure variables}

\item{diagonal}{a boolean indicating the diagonal structure of \eqn{C}}
}
\value{
the selected optimal \eqn{\lambda}
}
\description{
Use cross-validation to select \eqn{\lambda} for estimating \eqn{\Omega}.
Split the data into two parts and estimate \eqn{C} on both data sets. Then, for each
\eqn{lambda}, calculate \eqn{\Omega} on the first dataset and calculate the loss on the second dataset.
Find the lambda which minimizes \eqn{C \cdot \Omega - log(|\Omega|)}
}
