% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvLambda.R
\name{cvLambda}
\alias{cvLambda}
\title{Use cross-validation to select \eqn{\lambda} for estimating \eqn{\Omega}.
Split the data into two parts and estimate \eqn{C} on both data sets. Then, for each
\eqn{lambda}, calculate \eqn{\Omega} on the first dataset and calculate the loss on the second dataset.
Find the lambda which minimizes \eqn{C \cdot \Omega - log(|\Omega|)}}
\usage{
cvLambda(x, lambdas, AI, pure_vec, diagonal, k)
}
\arguments{
\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{lambdas}{a vector of numerical constants over which to search for the optimal \eqn{\lambda}}

\item{AI}{the estimated matrix \eqn{A_I} of dimensions \eqn{p \times K}}

\item{diagonal}{a boolean indicating the diagonal structure of \eqn{C}}

\item{k}{number of folds for \eqn{k}-fold cross-validation}

\item{pureVec}{a vector of indices of pure variables}
}
\value{
the selected optimal \eqn{\lambda}
}
\description{
Use cross-validation to select \eqn{\lambda} for estimating \eqn{\Omega}.
Split the data into two parts and estimate \eqn{C} on both data sets. Then, for each
\eqn{lambda}, calculate \eqn{\Omega} on the first dataset and calculate the loss on the second dataset.
Find the lambda which minimizes \eqn{C \cdot \Omega - log(|\Omega|)}
}
