% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dantzig.R
\name{Dantzig}
\alias{Dantzig}
\title{I am not really sure what this does.}
\usage{
Dantzig(C_hat, y, lbd)
}
\arguments{
\item{C_hat}{a matrix}

\item{y}{response vector of dimension ???}

\item{lbd}{\eqn{\lambda}, a positive constant}
}
\value{
???
}
\description{
I am not really sure what this does.
}
