% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devA.R
\name{devA}
\alias{devA}
\title{Find Matrix Deviation}
\usage{
devA(mat_A, mat_B)
}
\arguments{
\item{mat_A}{a matrix}

\item{mat_B}{a matrix of same dimensions as \code{mat_A}}
}
\value{
mean absolute deviation
}
\description{
Calculate the mean element-wise absolute deviation between two matrices.
}
