% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER.R
\name{ER}
\alias{ER}
\title{Run Essential Regression.}
\usage{
ER(
  Y,
  X,
  sigma,
  delta,
  beta_est = "NULL",
  CI = F,
  pred = T,
  lbd = 0.1,
  rep_CV = 50,
  diagonal = F,
  merge = F,
  equal_var = F,
  alpha_level = 0.05,
  support = NULL,
  correction = "Bonferroni",
  verbose = F
)
}
\arguments{
\item{Y}{a response vector of dimension \eqn{n}}

\item{X}{a data matrix of dimensions \eqn{n \times p}}

\item{sigma}{a sample correlation matrix of dimensions \eqn{p \times p}}

\item{delta}{\eqn{\delta}, a numerical constant used for thresholding}

\item{beta_est}{a string indicating the type of estimation to use for \eqn{\beta}}

\item{CI}{a boolean indicating whether to calculate confidence intervals for the \eqn{\beta} estimates}

\item{pred}{a boolean indicating whether to do prediction}

\item{lbd}{\eqn{\lambda}, a numerical constant used in thresholding}

\item{rep_CV}{number of replicates for cross-validation}

\item{diagonal}{a boolean indicating the diagonal structure of the data ???}

\item{merge}{a boolean indicating the merge type}

\item{equal_var}{a boolean indicating whether there is equal variance ??}

\item{alpha_level}{\eqn{\alpha}, a numerical constant used in confidence interval calculation}

\item{support}{a boolean ???}

\item{correction}{a string indicating the type of multi-testing correction to perform}

\item{verbose}{a boolean indicating whether to include printing}
}
\value{
a list of results from the Essential Regression framework including: \eqn{K = } number of clusters,
\eqn{\hat{A}}, \eqn{\hat{C}}, \eqn{\hat{I}}, the indices of the pure variables, \eqn{\hat{\Gamma}},
\eqn{\hat{\beta}}, \eqn{\alpha}-level confidence intervals (if requested), prediction results (if requested),
the optimal value of \eqn{\lambda} determined by cross-validation, the optimal value of \eqn{\delta}
determined by cross-validation, \eqn{Q}, and the variances of \eqn{\hat{\beta}}
}
\description{
Run Essential Regression.
}
