% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essregCV.R
\name{essregCV}
\alias{essregCV}
\title{Cross-Validation For Essential Regression}
\usage{
essregCV(
  k = 5,
  y,
  x,
  priors = NULL,
  delta,
  thresh_fdr = 0.2,
  lambda = 0.1,
  rep_cv = 50,
  alpha_level = 0.05,
  thresh = 0.001,
  perm_option = NULL,
  beta_est = "NULL",
  sel_corr = T,
  svm = F,
  diagonal = F,
  merge = F,
  equal_var = F,
  support = NULL,
  correction = T,
  change_all = F,
  verbose = F,
  delta_grid = NULL
)
}
\arguments{
\item{k}{an integer for number of folds to use in cross-validation}

\item{y}{a response vector of dimension \eqn{n}}

\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{priors}{a vector of names or indices of important features in data matrix if
using \link{priorER} or \code{NULL} if using \link{plainER}}

\item{delta}{\eqn{\delta}, a numerical constant used for thresholding}

\item{thresh_fdr}{a numerical constant used for thresholding the correlation matrix to
control the false discovery rate, default is 0.2}

\item{lambda}{\eqn{\lambda}, a numerical constant used in thresholding}

\item{rep_cv}{number of replicates for cross-validation}

\item{alpha_level}{\eqn{\alpha}, a numerical constant used in confidence interval calculation}

\item{thresh}{a numerical constant used as the threshold for convergence in Variational Bayes}

\item{perm_option}{a string indicating the type of permutation type do perform
(can be NULL, "x", "x_y", or "y_before_split")}

\item{beta_est}{a string indicating the type of estimation to use for \eqn{\beta}}

\item{sel_corr}{a boolean flag indicating whether to perform cross-validation by evaluating the correlation
between the predicted and true values of \eqn{y} (\code{T}) or by evaluating the prediction error via mse or auc (\code{F})}

\item{svm}{a boolean flag indicating whether to fit svm/svr and use that for prediction}

\item{diagonal}{a boolean indicating the diagonal structure of the data ???}

\item{merge}{a boolean indicating the merge type}

\item{equal_var}{a boolean indicating whether there is equal variance ??}

\item{support}{a boolean ???}

\item{correction}{a boolean flag indicating whether to perform Bonferroni multiple testing correction}

\item{change_all}{a boolean indicating whether to change all entries in \eqn{\hat{\Sigma}}
for an important feature (T) or to just change to more extreme values (F)}

\item{verbose}{a boolean indicating whether to include printing}
}
\value{
An object of class \sQuote{data.frame}
}
\description{
Perform k-fold cross-validation for essential regression to select
}
