% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_beta_dz.R
\name{Est_beta_dz}
\alias{Est_beta_dz}
\title{Estimate the values of the coefficients \eqn{\beta} with a Dantzig-type estimator.}
\usage{
Est_beta_dz(Y, X, A_hat, C_hat, I_hat, delta_opt, mu = 0.5, lbd = 0.5)
}
\arguments{
\item{Y}{response vector of dimension \eqn{p}}

\item{X}{data matrix of dimensions \eqn{n \times p}}

\item{A_hat}{a matrix of dimensions \eqn{p \times K}}

\item{C_hat}{estimated matrix}

\item{I_hat}{estimated matrix}

\item{delta_opt}{numeric value for \eqn{\delta}}

\item{mu}{???? \eqn{\mu}}

\item{lbd}{numeric value for \eqn{\lambda} used in soft thresholding}
}
\value{
Dantzig-type estimator results for \eqn{\beta}
}
\description{
Estimate the values of the coefficients \eqn{\beta} with a Dantzig-type estimator.
}
