% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_sigma2.R
\name{Est_sigma2}
\alias{Est_sigma2}
\title{I am not sure what this does.}
\usage{
Est_sigma2(Y, h, beta_est, C_hat)
}
\arguments{
\item{Y}{a response vector of dimension \eqn{p}}

\item{h}{Not sure}

\item{beta_est}{estimated values of \eqn{\beta}}

\item{C_hat}{estimated matrix}
}
\value{
???
}
\description{
I am not sure what this does.
}
