% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstAI.R
\name{EstAI}
\alias{EstAI}
\title{Use the given \eqn{\delta} to calculate the fitted \eqn{A_I}, pure variable
indices in list form and vector form. Also return estimated \eqn{Y} and \eqn{C} for
the following Dantzig estimation.}
\usage{
EstAI(Sigma, optDelta, se_est, merge)
}
\arguments{
\item{Sigma}{a correlation matrix of dimensions \eqn{p \times p}}

\item{optDelta}{\eqn{\delta}, a numerical constant}

\item{se_est}{???}

\item{merge}{a boolean indicating merge style}
}
\value{
a list including: \eqn{A_I}, a matrix of dimensions \eqn{p \times K},
a vector of the indices of estimated pure variables, and a list of the indices of
the estimated pure variables
}
\description{
Use the given \eqn{\delta} to calculate the fitted \eqn{A_I}, pure variable
indices in list form and vector form. Also return estimated \eqn{Y} and \eqn{C} for
the following Dantzig estimation.
}
