% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstAJDant.R
\name{EstAJDant}
\alias{EstAJDant}
\title{Estimate the matrix \eqn{A_J} with a Dantzig-type estimator.}
\usage{
EstAJDant(C_hat, Y_hat, lbd, se_est_J)
}
\arguments{
\item{C_hat}{a matrix}

\item{Y_hat}{a matrix}

\item{lbd}{\eqn{\lambda}, a positive constant}

\item{se_est_J}{the estimated standard errors of indices in \eqn{J}}
}
\value{
an estimate of matrix \eqn{A_J}
}
\description{
Estimate the matrix \eqn{A_J} with a Dantzig-type estimator.
}
