% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstAJInv.R
\name{EstAJInv}
\alias{EstAJInv}
\title{Estimate the matrix \eqn{A_J} using soft thresholding.}
\usage{
EstAJInv(Omega, Y, lbd)
}
\arguments{
\item{Omega}{an estimate of \eqn{C^{-1}} of dimensions \eqn{p \times p}}

\item{Y}{response matrix of dimensions \eqn{K \times |J|}}

\item{lbd}{\eqn{\lambda}, the chosen constant for the RHS constraint for soft thresholding}
}
\value{
a matrix of dimensions \eqn{|J| \times K}
}
\description{
Estimate the matrix \eqn{A_J} using soft thresholding.
}
