% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estBeta.R
\name{estBeta}
\alias{estBeta}
\title{Estimate \eqn{\beta}}
\usage{
estBeta(
  y,
  x,
  sigma,
  A_hat,
  C_hat,
  Gamma_hat,
  I_hat,
  I_hat_list,
  conf_int = T,
  alpha_level = 0.05,
  correction = TRUE,
  support = NULL
)
}
\arguments{
\item{y}{vector of responses of dimension \eqn{p}}

\item{x}{data matrix of dimensions \eqn{n \times p}}

\item{sigma}{sample correlation matrix of the data, \eqn{X}}

\item{A_hat}{estimated matrix of dimensions \eqn{p \times K}}

\item{C_hat}{estimated matrix}

\item{Gamma_hat}{estimated matrix}

\item{I_hat}{estimated matrix}

\item{I_hat_list}{estimated matrix}

\item{conf_int}{boolean indicated whether to calculate confidence intervals}

\item{alpha_level}{value for confidence intervals}

\item{correction}{a boolean flag indicating whether to perform Bonferroni multiple testing correction}

\item{support}{whether to do calculations on support of WHAT}
}
\value{
a list including the estimates for \eqn{\beta} and the confidence intervals (if requested)
}
\description{
Estimate the values of the coefficients \eqn{\beta} as well as confidence intervals for the estimates
if requested.
}
