% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estC.R
\name{estC}
\alias{estC}
\title{}
\usage{
estC(sigma, AI, diagonal = F)
}
\arguments{
\item{sigma}{a sample correlation matrix of dimensions \eqn{p \times p}}

\item{AI}{a matrix of dimensions \eqn{p \times K}}

\item{diagonal}{boolean indicating whether to only estimated diagonal elements of \eqn{C} (default = F)}
}
\value{
an estimate of \eqn{C} of dimensions \eqn{K \times K}
}
\description{

}
