% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estOmega.R
\name{estOmega}
\alias{estOmega}
\title{For a given \eqn{\lambda} and \eqn{C}, find \eqn{C^{-1}}.}
\usage{
estOmega(lambda, C)
}
\arguments{
\item{C}{a square, symmetric matrix}

\item{lbd}{\eqn{\lambda}}
}
\value{
\eqn{\Omega = C^{-1}}
}
\description{
For a given \eqn{\lambda} and \eqn{C}, find \eqn{C^{-1}}.
}
