% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstY.R
\name{EstY}
\alias{EstY}
\title{Estimate the matrix, \eqn{\Sigma_{\hat{TJ}}}.}
\usage{
EstY(Sigma, AI, pureVec)
}
\arguments{
\item{Sigma}{a sample correlation matrix of dimensions \eqn{p \times p}}

\item{AI}{a matrix of dimensions \eqn{p \times K}}

\item{pureVec}{a vector of indices of the pure nodes}
}
\value{
an estimate for \eqn{\Sigma_{\hat{TJ}}} of dimensions \eqn{K \times |J|}
}
\description{
Estimate the matrix, \eqn{\Sigma_{\hat{TJ}}}.
}
