% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClustMax.R
\name{findClustMax}
\alias{findClustMax}
\title{Find Clusters Maximums}
\usage{
findClustMax(sigma, cluster, clust_targ)
}
\arguments{
\item{sigma}{a correlation matrix of dimensions \eqn{p \times p}}

\item{cluster}{a cluster found by Essential Regression (list format)}

\item{clust_targ}{a vector of the target nodes in the cluster}
}
\value{
the absolute maximum correlation found out of all of the rows provided
}
\description{
Find the absolute minimum of the maximum correlations of each of the target nodes in the
given cluster.
}
