% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClustMed.R
\name{findClustMed}
\alias{findClustMed}
\title{Find the median correlation of a cluster by using the nodes in cluster as a list of
rows and searching over these rows for the median value in the sample correlation matrix.}
\usage{
findClustMed(sigma, cluster, clust_targ)
}
\arguments{
\item{sigma}{a correlation matrix of dimensions \eqn{p \times p}}

\item{cluster}{a cluster found by Essential Regression (list format)}

\item{clust_pure}{a vector of the pure nodes in the cluster}
}
\value{
the median correlation found out of all of the rows provided
}
\description{
Find the median correlation of a cluster by using the nodes in cluster as a list of
rows and searching over these rows for the median value in the sample correlation matrix.
}
