% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRho.R
\name{findRho}
\alias{findRho}
\title{Calculate the optimal value of \eqn{\rho} to use in \eqn{\Delta} that maximizes the marginal log-likelihood of the data, \eqn{x}, for fixed \eqn{\alpha}.}
\usage{
findRho(x, imp, alpha, rho_range)
}
\arguments{
\item{x}{data matrix of dimensions \eqn{n \times p}}

\item{imp}{a vector of indices indicating important features}

\item{alpha}{\eqn{\alpha}, a positive value between 0 and 1}

\item{rho_range}{a range of values to try for \eqn{\rho} (should be some subset of \[-1, 1\])}
}
\value{
a list containing the marginal log-likelihoods, the optimal value of \eqn{\rho},
the adjusted correlation matrix, and a plot of the marginal log-likelihoods
}
\description{
Calculate the optimal value of \eqn{\rho} to use in \eqn{\Delta} that maximizes the marginal log-likelihood of the data, \eqn{x}, for fixed \eqn{\alpha}.
}
