% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRowMax.R
\name{FindRowMax}
\alias{FindRowMax}
\title{Calculate the maximal absolute value for each row of the given matrix.}
\usage{
FindRowMax(Sigma)
}
\arguments{
\item{Sigma}{a matrix of dimensions \eqn{p \times p}}
}
\value{
a list including the indices of the maximal absolute values of the rows of
\eqn{\Sigma} as well as the values
}
\description{
Calculate the maximal absolute value for each row of the given matrix.
}
