% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRowMaxInd.R
\name{FindRowMaxInd}
\alias{FindRowMaxInd}
\title{Calculate indices of each row such that the absolute values of these indices
are within \eqn{2\delta} of the maximal absolute value \eqn{M} of this row.}
\usage{
FindRowMaxInd(i, M, arg_M, vector, delta, se_est)
}
\arguments{
\item{i}{the row index}

\item{M}{the maximal absolute value of row \eqn{i}}

\item{arg_M}{the indices at which \eqn{M} is achieved in the row}

\item{vector}{the row}

\item{delta}{\eqn{\delta}, a numeric constant}

\item{se_est}{???}
}
\value{
a vector of indices
}
\description{
Calculate indices of each row such that the absolute values of these indices
are within \eqn{2\delta} of the maximal absolute value \eqn{M} of this row.
}
