% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindSignPureNode.R
\name{FindSignPureNode}
\alias{FindSignPureNode}
\title{Estimate the sign subpartition of pure node sets. If there is an element
of a list is empty, then a empty list will be put in that position.}
\usage{
FindSignPureNode(pureList, Sigma)
}
\arguments{
\item{pureList}{a list of pure node indices}

\item{Sigma}{a sample correlation matrix of dimensions \eqn{p \times p}}
}
\value{
a list of sign subpartitions of pure node indices
}
\description{
Estimate the sign subpartition of pure node sets. If there is an element
of a list is empty, then a empty list will be put in that position.
}
