% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussKOs.R
\name{gaussKOs}
\alias{gaussKOs}
\title{Gaussian Knockoffs}
\usage{
gaussKOs(
  z,
  y,
  statistic = knockoff::stat.glmnet_lambdasmax,
  fdr = 0.05,
  iters = 20
)
}
\arguments{
\item{z}{a matrix of dimensions \eqn{p \times K}}

\item{y}{a response vector of dimension \eqn{n}, standardized}

\item{statistic}{a statistic to use for variable selection}

\item{fdr}{the target false discovery rate}

\item{iters}{the number of iterations to perform}
}
\value{
a vector of selected variable indices
}
\description{
A feature selection method using Model X Knockoffs
}
