% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErrors.R
\name{getErrors}
\alias{getErrors}
\title{Calculate the error rates for an estimated matrix.}
\usage{
getErrors(estA, trueA, flagPerm = TRUE)
}
\arguments{
\item{estA}{an estimated matrix}

\item{trueA}{the true matrix of same dimensions as \code{estA}}

\item{flagPerm}{boolean for whether to use the optimal permuted version of \code{estA} for the calculations}
}
\value{
the error rates for \code{estA}: specificity, sensitivity, false negative rate, and false positive rate
}
\description{
Calculate the error rates for an estimated matrix.
}
