% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFreq.R
\name{getFreq}
\alias{getFreq}
\title{Fin Mixed Variable Frequencies}
\usage{
getFreq(er_res)
}
\arguments{
\item{er_res}{an object return by \link[ER]{ER()}}
}
\value{
a matrix with the same number of rows as there are mixed variables in \code{er_res} and with
a column containing frequencies as well as a column for each cluster that contain a 1 if that variable
is found in that cluster and a 0 if not.
}
\description{
Find the frequency with which each mixed variable appears in the results of Essential Regression and
return a table of the frequencies as well as the clusters in which each variable appears.
}
