% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierER.R
\name{hierER}
\alias{hierER}
\title{Hierarchical Essential Regression.}
\usage{
hierER(
  y,
  x,
  delta,
  thresh_fdr = 0.2,
  beta_est = "LS",
  conf_int = T,
  pred = T,
  lambda = 0.1,
  rep_cv = 50,
  diagonal = F,
  merge = F,
  equal_var = F,
  alpha_level = 0.05,
  support = NULL,
  correction = T,
  sig_betas = NULL,
  niter = 20
)
}
\arguments{
\item{y}{a response vector of dimension \eqn{n}, must be continous}

\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{delta}{\eqn{\delta}, a numerical constant used for thresholding}

\item{thresh_fdr}{a numerical constant used for thresholding the correlation matrix to
control the false discovery rate, default is 0.2}

\item{beta_est}{a string indicating the type of estimation to use for \eqn{\beta}}

\item{conf_int}{a boolean indicating whether to calculate confidence intervals for the \eqn{\beta} estimates}

\item{pred}{a boolean indicating whether to do prediction}

\item{lambda}{\eqn{\lambda}, a numerical constant used in thresholding}

\item{rep_cv}{number of replicates for cross-validation}

\item{diagonal}{a boolean indicating the diagonal structure of the data ???}

\item{merge}{a boolean indicating the merge type}

\item{equal_var}{a boolean indicating whether there is equal variance ??}

\item{alpha_level}{\eqn{\alpha}, a numerical constant used in confidence interval calculation}

\item{support}{a boolean ???}

\item{correction}{a boolean flag indicating whether to perform Bonferroni multiple testing correction}

\item{sig_betas}{either NULL or the proportion of \eqn{\beta}s to use when determining significant components
of \eqn{\mathbf{Z}}}

\item{niter}{the number of iterations to perform in the variable selection step}
}
\value{
a list containing the results of the intial run of `plainER()`, the newly estimated \eqn{\beta},
the predicted values using the new \eqn{\beta}, and a list containing the results of variable
selection via Model X knockoffs
}
\description{
Essential Regression with interaction terms among the components of \eqn{\mathbf{Z}}.
}
