% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVS.R
\name{IVS}
\alias{IVS}
\title{Iterative Variable Selection}
\usage{
IVS(y, Z, verbose = F)
}
\arguments{
\item{y}{a response vector of dimension \eqn{n}}

\item{Z}{a matrix of dimensions \eqn{p \times K}}
}
\value{
a vector of selected variable indices
}
\description{
A feature selection method suitable for variables with high correlation
}
