% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVSRobust.R
\name{IVSRobust}
\alias{IVSRobust}
\title{Iterative Variable Selection - Robustified}
\usage{
IVSRobust(y, z, freq_tab)
}
\arguments{
\item{y}{a response vector of dimension \eqn{n}}

\item{z}{a matrix of dimensions \eqn{p \times K}}

\item{priors}{a vector of indices indicating the important features}

\item{er_res}{an object returned by \code{plainER} or \code{priorER}}
}
\value{
a vector of selected variable indices
}
\description{
A feature selection method suitable for variables with high correlation
}
