% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logGamma.R
\name{logGamma}
\alias{logGamma}
\title{Calculate natural logarithm of multivariate Gamma function for evaluating the
marginal log-likelihood of \eqn{x}. This function both adds and subtracts values
as in the log-likelihood of \eqn{x}. See \code{\link{logLik}} for more details.}
\usage{
logGamma(p, val1, val2)
}
\arguments{
\item{p}{for p-adic Gamma function}

\item{val1}{first value}

\item{val2}{second value}
}
\value{
the natural logarithm of multivariate Gamma function
}
\description{
Calculate natural logarithm of multivariate Gamma function for evaluating the
marginal log-likelihood of \eqn{x}. This function both adds and subtracts values
as in the log-likelihood of \eqn{x}. See \code{\link{logLik}} for more details.
}
