% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\title{Evaluate the marginal log-likelihood of the data, \eqn{x}, given weight \eqn{\alpha} and expert matrix \eqn{\Delta}:
ADD EQUATION}
\usage{
logLik(alpha, Delta, x)
}
\arguments{
\item{alpha}{weight for the intensity of the balance between the expert knowledge matrix, \eqn{\Delta}, and the data}

\item{Delta}{the correlation matrix of dimensions \eqn{p \times p} created with regard to expert knowledge}

\item{x}{data matrix of dimensions \eqn{n \times p}}
}
\value{
numeric
}
\description{
Evaluate the marginal log-likelihood of the data, \eqn{x}, given weight \eqn{\alpha} and expert matrix \eqn{\Delta}:
ADD EQUATION
}
