% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDelta.R
\name{makeDelta}
\alias{makeDelta}
\title{Construct the expert knowledge matrix used as the target for the sample correlation matrix
reweighting. This is the first step of including prior knowledge in Essential Regression.}
\usage{
makeDelta(
  y,
  x,
  imp,
  delta_range,
  lbd_range,
  beta_est = "LS",
  diagonal = F,
  merge = F,
  equal_var = F,
  alpha_level = 0.05,
  support = F,
  correction = "Bonferroni"
)
}
\arguments{
\item{y}{a response vector of dimension \eqn{n}}

\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{imp}{a vector of the indices of the features determined to be important by expert knowledge}

\item{delta_range}{a vector of values to search over when picking the optimal \eqn{\delta} by cross-validation}

\item{lbd_range}{a vector of values to search over when picking the optimal \eqn{\lambda} by cross-validation}
}
\value{

}
\description{
Construct the expert knowledge matrix used as the target for the sample correlation matrix
reweighting. This is the first step of including prior knowledge in Essential Regression.
}
