% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDelta2.R
\name{makeDelta2}
\alias{makeDelta2}
\title{Construct the expert knowledge matrix used as the target for the sample correlation matrix
reweighting. This is the first step of including prior knowledge in Essential Regression.
The column and row in the sample correlation matrix corresponding to the important feature
are adjusted in the following way:}
\usage{
makeDelta2(y, x, imp, er_res, change_all = F, equal_var = F)
}
\arguments{
\item{y}{a response vector of dimension \eqn{n}}

\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{imp}{the index of the important feature}

\item{er_res}{the output of a run of \code{\link{ER}}}

\item{equal_var}{a boolean indicating whether the columns of \code{x} are assumed to have equal variance}
}
\value{
the expert knowledge matrix, \eqn{\Delta}, of dimensions \eqn{p \times p}
}
\description{
First, find the absolute maximum entry in the sample correlation matrix (excluding diagonal elements)
Then calculate the cutoff/threshold value used in determining node purity by \code{\link{FindRowMaxInd()}}
Next, calculate the values used in the replacement vector by subtracting the cutoff/threshold value from the
absolute maximum entry for each row. We then get a vector of values used to replace each of \eqn{\hat{\Sigma}_{imp, i}}
and \eqn{\hat{\Sigma}_{i, imp}}. If the \eqn{i}th replacement value in the replacement vector is greater in absolute
value than the value already found at \eqn{\hat{\Sigma}_{i, imp}} and \eqn{\hat{\Sigma}_{imp, i}}, then replace
that entry in \eqn{\hat{\Sigma}} with the replacement value. Otherwise, leave the entry alone.
Finally, we find the nearest positive definite matrix to our new \eqn{\hat{\Sigma}} with a diagonal of 1s.
}
