% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDelta_old.R
\name{makeDelta_old}
\alias{makeDelta_old}
\title{Construct the expert knowledge matrix used as the target for the sample correlation matrix
reweighting. This is the first step of including prior knowledge in Essential Regression.}
\usage{
makeDelta_old(y, x, imp, er_res, type = "med", as_names = T)
}
\arguments{
\item{y}{a response vector of dimension \eqn{n}}

\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{imp}{a vector of the features determined to be important by expert knowledge}

\item{er_res}{the output of a run of \code{\link{ER}}}

\item{type}{string indicating what type of replacement calculation to do}

\item{as_names}{boolean indicating whether the important features are reported as a vector of
indices (integers) or a vector of feature names (strings)}
}
\value{
a matrix resulting from replacing the low correlations with the median or minimal value
in the given row/column. returns NULL if all important features in \code{imp} are already present
in the ER results
}
\description{
Construct the expert knowledge matrix used as the target for the sample correlation matrix
reweighting. This is the first step of including prior knowledge in Essential Regression.
}
