% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePosDef.R
\name{makePosDef}
\alias{makePosDef}
\title{Make Matrix Positive Definite}
\usage{
makePosDef(samp_corr)
}
\arguments{
\item{samp_corr}{a sample correlation matrix of dimension \eqn{p \times p}}
}
\value{
a positive definite matrix of the same dimension as \code{samp_corr}
}
\description{
Transform the given matrix so that it is positive definite.
}
