% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{Merge}
\alias{Merge}
\title{Merge a group with other groups containing common nodes.}
\usage{
Merge(groupList, groupVec)
}
\arguments{
\item{groupList}{a list of groups of node indices}

\item{groupVec}{a vector of node indices}
}
\value{
a list of the merged results
}
\description{
Merge a group with other groups containing common nodes.
}
