% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_union.R
\name{Merge_union}
\alias{Merge_union}
\title{Merge the provided vector of nodes into the list of vectors of nodes but use union rather than the intersection.}
\usage{
Merge_union(groupList, groupVec)
}
\arguments{
\item{groupList}{a list of groups of node indices}

\item{groupVec}{a vector of node indices}
}
\value{
a list of the merged results
}
\description{
Merge the provided vector of nodes into the list of vectors of nodes but use union rather than the intersection.
}
