% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mse.R
\name{Mse}
\alias{Mse}
\title{Calculate the element-wise mean squared error of a matrix.}
\usage{
Mse(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix of same dimensions as \code{A}}
}
\value{
MSE
}
\description{
Calculate the element-wise mean squared error of a matrix.
}
