% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offSum.R
\name{offSum}
\alias{offSum}
\title{Calculate the sum of squares of the upper off-diagonal elements of two matrices}
\usage{
offSum(M, N, weights)
}
\arguments{
\item{M}{a matrix}

\item{N}{a matrix of same dimensions as \code{M}}

\item{weights}{the weights to use in calculating the sum}
}
\value{
the sum of squares
}
\description{
Calculate the sum of squares of the upper off-diagonal elements of two matrices
}
