% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRun.R
\name{parseRun}
\alias{parseRun}
\title{Parse .yaml file for Essential Regression}
\usage{
parseRun(yaml_path, delta, nreps)
}
\arguments{
\item{yaml_path}{a string path to the .yaml file}

\item{delta}{\eqn{\delta}, a numeric constant or vector of constants used in
thresholding during node purity testing}

\item{nreps}{number of replicates to do for cross-validation}
}
\value{
none
}
\description{
Read a .yaml file, parse arguments, and run Essential Regression.
}
