% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pred_Z_BLP.R
\name{Pred_Z_BLP}
\alias{Pred_Z_BLP}
\title{Additional prediction function using the support of \eqn{\beta}.}
\usage{
Pred_Z_BLP(X, A_hat, C_hat, est_Gamma, S_beta)
}
\arguments{
\item{X}{data matrix of dimensions \eqn{n \times p}}

\item{A_hat}{a matrix of dimensions \eqn{p \times K}}

\item{C_hat}{a matrix of dimensions}

\item{est_Gamma}{a matrix of dimensions}

\item{S_beta}{????}
}
\value{
????
}
\description{
Additional prediction function using the support of \eqn{\beta}.
}
