% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pred_Z_BLP_avg.R
\name{Pred_Z_BLP_avg}
\alias{Pred_Z_BLP_avg}
\title{An additional prediction function that uses the support of \eqn{\beta}.}
\usage{
Pred_Z_BLP_avg(X, A_hat, C_hat, S_beta)
}
\arguments{
\item{X}{a data matrix of dimensions \eqn{n \times p}}

\item{A_hat}{a matrix of dimensions \eqn{p \times K}}

\item{C_hat}{a matrix of dimensions}

\item{S_beta}{???}
}
\value{
????
}
\description{
An additional prediction function that uses the support of \eqn{\beta}.
}
