% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{Prediction}
\alias{Prediction}
\title{Perform prediction on the data using the ER predictor.}
\usage{
Prediction(Y, X, sigma, A_hat, Gamma_hat, I_hat)
}
\arguments{
\item{Y}{a response vector of dimension \eqn{n}}

\item{X}{a data matrix of dimensions \eqn{n \times p}}

\item{sigma}{a correlation matrix of dimensions \eqn{p \times p}}

\item{A_hat}{a matrix of dimensions \eqn{p \times K}}

\item{Gamma_hat}{a matrix of dimensions}

\item{I_hat}{a matrix of dimensions}
}
\value{
a list including \eqn{\hat{\theta}}, the predicted values, and \eqn{\Theta}
}
\description{
Perform prediction on the data using the ER predictor.
}
