% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predZAvg.R
\name{predZAvg}
\alias{predZAvg}
\title{Estimate \eqn{Z} (Average)}
\usage{
predZAvg(x, A_hat, C_hat, S_beta)
}
\arguments{
\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{A_hat}{a matrix of dimensions \eqn{p \times K}}

\item{C_hat}{a matrix of dimensions}

\item{S_beta}{???}
}
\value{
????
}
\description{
An additional prediction function that uses the support of \eqn{\beta}.
}
