% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorER.R
\name{priorER}
\alias{priorER}
\title{Prior Information Essential Regression}
\usage{
priorER(
  y,
  x,
  imps,
  sigma = NULL,
  delta,
  thresh_fdr = 0.2,
  beta_est = "NULL",
  conf_int = F,
  pred = T,
  lambda = 0.1,
  rep_cv = 50,
  diagonal = F,
  merge = F,
  equal_var = F,
  alpha_level = 0.05,
  thresh = 0.001,
  support = NULL,
  correction = T,
  change_all = F,
  verbose = F
)
}
\arguments{
\item{y}{a response vector of dimension \eqn{n}}

\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{imps}{a vector of important feature indices/names}

\item{sigma}{a sample correlation matrix of dimensions \eqn{p \times p}}

\item{delta}{\eqn{\delta}, a numerical constant used for thresholding}

\item{thresh_fdr}{a numerical constant used for thresholding the correlation matrix to
control the false discovery rate, default is 0.2}

\item{beta_est}{a string indicating the type of estimation to use for \eqn{\beta}}

\item{conf_int}{a boolean indicating whether to calculate confidence intervals for the \eqn{\beta} estimates}

\item{pred}{a boolean indicating whether to do prediction}

\item{lambda}{\eqn{\lambda}, a numerical constant used in thresholding}

\item{rep_cv}{number of replicates for cross-validation}

\item{diagonal}{a boolean indicating the diagonal structure of the data ???}

\item{merge}{a boolean indicating the merge type}

\item{equal_var}{a boolean indicating whether there is equal variance ??}

\item{alpha_level}{\eqn{\alpha}, a numerical constant used in confidence interval calculation}

\item{thresh}{a numerical constant used as the threshold for convergence in Variational Bayes}

\item{support}{a boolean ???}

\item{correction}{a boolean flag indicating whether to perform Bonferroni multiple testing correction}

\item{change_all}{a boolean indicating whether to change all entries in \eqn{\hat{\Sigma}}
for an important feature (T) or to just change to more extreme values (F)}

\item{verbose}{a boolean indicating whether to include printing}
}
\value{
a list of results from the Essential Regression framework including: \eqn{K = } number of clusters,
\eqn{\hat{A}}, \eqn{\hat{C}}, \eqn{\hat{I}}, the indices of the pure variables, \eqn{\hat{\Gamma}},
\eqn{\hat{\beta}}, \eqn{\alpha}-level confidence intervals (if requested), prediction results (if requested),
the optimal value of \eqn{\lambda} determined by cross-validation, the optimal value of \eqn{\delta}
determined by cross-validation, \eqn{Q}, and the variances of \eqn{\hat{\beta}}
}
\description{
Run Essential Regression with prior information.
}
