% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pureRowInd.R
\name{pureRowInd}
\alias{pureRowInd}
\title{For given matrix \eqn{A}, find the row indices which correspond to pure nodes.}
\usage{
pureRowInd(A)
}
\arguments{
\item{A}{a matrix of dimensions \eqn{p \times K}}
}
\value{
a vector of indices
}
\description{
For given matrix \eqn{A}, find the row indices which correspond to pure nodes.
}
