% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readER.R
\name{readER}
\alias{readER}
\title{Function to read all, pure, and mixed variables from ER results.}
\usage{
readER(er_res)
}
\arguments{
\item{er_res}{a list output by the function \code{\link{ER}}}
}
\value{
a list including the clusters, pure variables, and mixed varlues
found by ER
}
\description{
Function to read all, pure, and mixed variables from ER results.
}
