% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecoverAI.R
\name{RecoverAI}
\alias{RecoverAI}
\title{Recover the estimated submatrix \eqn{A_I} by given the pure node group.}
\usage{
RecoverAI(estGroupList, p)
}
\arguments{
\item{estGroupList}{a list of group indices of the pure nodes with sign}

\item{p}{the number of features}
}
\value{
a matrix of dimensions \eqn{p \times K}
}
\description{
Recover the estimated submatrix \eqn{A_I} by given the pure node group.
}
