% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siNetwork.R
\name{siNetwork}
\alias{siNetwork}
\title{Generate Node Network}
\usage{
siNetwork(x, sigma, er_res, clust_num, equal_var = F, merge = F)
}
\arguments{
\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{sigma}{a sample correlation matrix of dimensions \eqn{p \times p}}

\item{er_res}{the results from running Essential Regression with either
\link[priorER]{priorER()} or \link[plainER]{plainER()}}

\item{clust_num}{an integer indicating the cluster number}

\item{equal_var}{a boolean flag indicating whether the columns of \code{x} have equal variance}

\item{merge}{a boolean flag indicating merge type}
}
\description{
Create an interactive node network of correlation structure.
}
