% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleton.R
\name{singleton}
\alias{singleton}
\title{Check if there exists an element of the given list has length equal to 1}
\usage{
singleton(estPureIndices)
}
\arguments{
\item{estPureIndices}{estimated indices of pure nodes}
}
\value{
true or false depending upon existence of singleton element
}
\description{
Check if there exists an element of the given list has length equal to 1
}
