% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_row.R
\name{solve_row}
\alias{solve_row}
\title{???}
\usage{
solve_row(col_ind, C, lbd)
}
\arguments{
\item{col_ind}{indices of columns}

\item{C}{matrix?}

\item{lbd}{\eqn{\labmda}}
}
\value{
value
}
\description{
???
}
