% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestPure.R
\name{TestPure}
\alias{TestPure}
\title{For given row, check if it is a pure node by iteratively checking the nodes
in \code{Si}. Return TRUE if the given row corresponds to a pure variable.}
\usage{
TestPure(Sigma_row, rowInd, Si, Ms, arg_Ms, delta, se_est)
}
\arguments{
\item{rowInd}{the index of the row}

\item{Si}{a vector of indices}

\item{Ms}{a vector of the largest absolute values of each of the rows in \code{Sigma}}

\item{delta}{\eqn{\delta}, a numerical constant}

\item{Sigma}{a matrix of dimensions \eqn{p \times p}}
}
\value{
TRUE or FALSE
}
\description{
For given row, check if it is a pure node by iteratively checking the nodes
in \code{Si}. Return TRUE if the given row corresponds to a pure variable.
}
