% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toAlpha.R
\name{toAlpha}
\alias{toAlpha}
\title{Transform \eqn{\Phi} to \eqn{\alpha} for log-likelihood calculation}
\usage{
toAlpha(phi, n, p)
}
\arguments{
\item{phi}{numeric value}

\item{n}{number of rows/samples in data matrix (integer)}

\item{p}{number of columns/features in data matrix (integer)}
}
\value{
numeric
}
\description{
Transform \eqn{\Phi} to \eqn{\alpha} for log-likelihood calculation
}
