% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPhi.R
\name{toPhi}
\alias{toPhi}
\title{Transform \eqn{\alpha} to \eqn{\Phi} for log-likelihood calculation}
\usage{
toPhi(alpha, n, p)
}
\arguments{
\item{alpha}{numeric value}

\item{n}{number of rows/samples in data matrix (integer)}

\item{p}{number of columns/features in data matrix (integer)}
}
\value{
numeric
}
\description{
Transform \eqn{\alpha} to \eqn{\Phi} for log-likelihood calculation
}
