% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transSigHat.R
\name{transSigHat}
\alias{transSigHat}
\title{Create the transformed version of \eqn{\hat{\Sigma}} by replaced rows/columns of \eqn{\hat{\Sigma}}, the
sample correlation matrix of the data, \eqn{x}, with a specified value, \eqn{r}}
\usage{
transSigHat(x, imp, r)
}
\arguments{
\item{x}{data matrix of dimensions \eqn{n \times p}}

\item{imp}{vector of indices of important features}

\item{r}{replacement value}
}
\value{
a transformed version of \eqn{\hat{\Sigma}}, the expert knowledge matrix for Essential Regression
}
\description{
Create the transformed version of \eqn{\hat{\Sigma}} by replaced rows/columns of \eqn{\hat{\Sigma}}, the
sample correlation matrix of the data, \eqn{x}, with a specified value, \eqn{r}
}
