% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_data.R
\docType{data}
\name{x_data}
\alias{x_data}
\title{Gene expression data from Var50.}
\format{
A data frame with 24 rows and 800 columns.
}
\usage{
x_data
}
\description{
Example data for the Essential Regression vignette.
}
\keyword{datasets}
