% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_data.R
\docType{data}
\name{y_data}
\alias{y_data}
\title{Response data from Var50.}
\format{
A data frame with 24 rows and 1 column.
}
\usage{
y_data
}
\description{
Example data for the Essential Regression vignette.
}
\keyword{datasets}
