% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesVarSel.R
\name{bayesVarSel}
\alias{bayesVarSel}
\title{An implementation of the variational algorithm for Bayesian
variable selection paper by Huang, Wang, Liang (2016) \link{https://arxiv.org/pdf/1602.07640.pdf}.}
\usage{
bayesVarSel(y, x, imps, er_res)
}
\arguments{
\item{y}{a vector of responses}

\item{er_res}{a list output by the function \code{\link{plainER}} or \code{\link{priorER}}}

\item{alpha}{a numeric constant for the p-value threshold}
}
\value{
a list including the clusters, pure variables, mixed variables, the significant
cluster indices, and the features found in the significant clusters
found by ER
}
\description{
An implementation of the variational algorithm for Bayesian
variable selection paper by Huang, Wang, Liang (2016) \link{https://arxiv.org/pdf/1602.07640.pdf}.
}
