% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesVarSel.R
\name{bayesVarSel}
\alias{bayesVarSel}
\title{Variational Bayes Variable Selection}
\usage{
bayesVarSel(x, y, imps, er_res, thresh = 0.01)
}
\arguments{
\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{y}{a response vector of dimension \eqn{n}}

\item{imps}{a vector of important feature names/indices}

\item{er_res}{a list output by the function link[EssReg]{plainER} or \link[EssReg]{priorER}}

\item{thresh}{a numeric constant used in determining convergence}
}
\value{
a list including the clusters, pure variables, mixed variables, the significant
cluster indices, and the features found in the significant clusters
found by ER
}
\description{
An implementation of the variational algorithm for Bayesian
variable selection paper by Huang, Wang, Liang (2016).
}
