% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaBMA.R
\name{betaBMA}
\alias{betaBMA}
\title{PriorER - Bayesian Model Averaging For \eqn{\beta}s}
\usage{
betaBMA(x, y, er_res, imps, estim = "HPM")
}
\arguments{
\item{x}{a data matrix of dimensions \eqn{n \times p}}

\item{y}{a response vector of dimension \eqn{n}}

\item{er_res}{the result of \link{plainER} or \link{priorER}}

\item{imps}{a vector of important variables}

\item{estim}{the type of estimator used for Bayesian Model Averaging. options include
"BMA" = Bayesian Model Averaging Model, "HPM" = Highest Probability Model,
"MPM" = Median Probability Model}
}
\value{
a vector of \eqn{\beta} estimates
}
\description{
Find new estimates for \eqn{\beta}s using Bayesian Adaptive
Sampling for Bayesian Model Averaging as in the \code{BAS} package by Clyde et al.
}
