% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estBetaDant.R
\name{estBetaDant}
\alias{estBetaDant}
\title{Estimate \eqn{\beta} (Dantzig)}
\usage{
estBetaDant(y, x, A_hat, C_hat, I_hat, delta, mu = 0.5, lambda = 0.5)
}
\arguments{
\item{y}{response vector of dimension \eqn{p}}

\item{x}{data matrix of dimensions \eqn{n \times p}}

\item{A_hat}{a matrix of dimensions \eqn{p \times K}}

\item{C_hat}{estimated matrix}

\item{I_hat}{estimated matrix}

\item{delta}{numeric value for \eqn{\delta}}

\item{mu}{???? \eqn{\mu}}

\item{lambda}{numeric value for \eqn{\lambda} used in soft thresholding}
}
\value{
Dantzig-type estimator results for \eqn{\beta}
}
\description{
Estimate the values of the coefficients \eqn{\beta} with a Dantzig-type estimator.
}
