% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstC.R
\name{EstC}
\alias{EstC}
\title{Estimate \eqn{C}. If diagonal=True, estimate only diagonal elements.}
\usage{
EstC(Sigma, AI, diagonal)
}
\arguments{
\item{Sigma}{a sample correlation matrix of dimensions \eqn{p \times p}}

\item{AI}{a matrix of dimensions \eqn{p \times K}}

\item{diagonal}{boolean indicating whether to only estimated diagonal elements of \eqn{C}}
}
\value{
an estimate of \eqn{C} of dimensions \eqn{K \times K}
}
\description{
Estimate \eqn{C}. If diagonal=True, estimate only diagonal elements.
}
