% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPureNode.R
\name{FindPureNode}
\alias{FindPureNode}
\title{Estimate list of pure node indices for given \eqn{\Sigma} and \eqn{\delta}.}
\usage{
FindPureNode(off_Sigma, delta, Ms, arg_Ms, se_est, merge)
}
\arguments{
\item{off_Sigma}{a sample correlation matrix of dimensions \eqn{p \times p}}

\item{delta}{\eqn{\delta}, a numerical constant}

\item{Ms}{the largest absolute values of each row of \code{off_Sigma}}

\item{arg_Ms}{??}

\item{se_est}{??}

\item{merge}{boolean indicating merge style}
}
\value{
a list including the list of estimated pure node indices and a vector
of the estimated pure node indices
}
\description{
Estimate list of pure node indices for given \eqn{\Sigma} and \eqn{\delta}.
}
