% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readER.R
\name{readER}
\alias{readER}
\title{Read Essential Regression Results}
\usage{
readER(er_res, y, alpha = 0.05)
}
\arguments{
\item{er_res}{a list output by the function \link[EssReg]{plainER} or \link[EssReg]{priorER}}

\item{y}{a vector of responses}

\item{alpha}{a numeric constant for the p-value threshold}
}
\value{
a list including the clusters, pure variables, mixed variables, the significant
cluster indices, and the features found in the significant clusters
found by ER
}
\description{
Function to read all, pure, and mixed variables from ER results.
}
